+++ react-client/vite.config.ts
@@ -1,10 +1,11 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 
-// https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
-  build:{
+  plugins: [
+    react()
+  ],
+  build: {
     outDir: '../src/main/external-resources/web/react-client',
     emptyOutDir: true,
     assetsDir: 'static',
@@ -24,8 +25,8 @@ export default defineConfig({
   },
   server: {
     open: '/',
-    proxy:{
+    proxy: {
       '/': 'http://localhost:9001',
-	}
+    }
   }
 })